+++ src/manifolds/GeneralUnitaryMatrices.jl
@@ -1,31 +1,31 @@
 @doc """
     AbstractMatrixType
 
-A plain type to distinguish different types of matrices, for example [`DeterminantOneMatrices`](@ref)
-and [`AbsoluteDeterminantOneMatrices`](@ref).
+A plain type to distinguish different types of matrices, for example [`DeterminantOneMatrixType`](@ref)
+and [`AbsoluteDeterminantOneMatrixType`](@ref).
 """
 abstract type AbstractMatrixType end
 
 @doc """
-    DeterminantOneMatrices <: AbstractMatrixType
+    DeterminantOneMatrixType <: AbstractMatrixType
 
 A type to indicate that we require special (orthogonal / unitary) matrices, i.e. of determinant 1.
 """
-struct DeterminantOneMatrices <: AbstractMatrixType end
+struct DeterminantOneMatrixType <: AbstractMatrixType end
 
 @doc """
-    AbsoluteDeterminantOneMatrices <: AbstractMatrixType
+    AbsoluteDeterminantOneMatrixType <: AbstractMatrixType
 
 A type to indicate that we require (orthogonal / unitary) matrices with normed determinant,
 i.e. that the absolute value of the determinant is 1.
 """
-struct AbsoluteDeterminantOneMatrices <: AbstractMatrixType end
+struct AbsoluteDeterminantOneMatrixType <: AbstractMatrixType end
 
 @doc raw"""
     GeneralUnitaryMatrices{T,𝔽,S<:AbstractMatrixType} <: AbstractDecoratorManifold
 
 A common parametric type for matrices with a unitary property of size ``n×n`` over the field ``𝔽``
-which additionally have the `AbstractMatrixType`, e.g. are `DeterminantOneMatrices`.
+which additionally have the `AbstractMatrixType`, e.g. are [`DeterminantOneMatrixType`](@ref).
 """
 struct GeneralUnitaryMatrices{T,𝔽,S<:AbstractMatrixType} <: AbstractDecoratorManifold{𝔽}
     size::T
@@ -56,7 +56,7 @@ i.e. that ``p`` has a determinant of absolute value one.
 The tolerance for the last test can be set using the `kwargs...`.
 """
 function check_point(
-    M::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrices},
+    M::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrixType},
     p;
     kwargs...,
 ) where {𝔽}
@@ -84,7 +84,7 @@ i.e. that ``p`` has a determinant of absolute value one, i.e. that ``p^{\mathrm{
 The tolerance for the last test can be set using the `kwargs...`.
 """
 function check_point(
-    M::GeneralUnitaryMatrices{<:Any,𝔽,DeterminantOneMatrices},
+    M::GeneralUnitaryMatrices{<:Any,𝔽,DeterminantOneMatrixType},
     p;
     kwargs...,
 ) where {𝔽}
@@ -661,7 +661,7 @@ Return the injectivity radius for general unitary matrix manifolds, which is[^1]
 injectivity_radius(::GeneralUnitaryMatrices) = π
 
 @doc raw"""
-    injectivity_radius(G::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+    injectivity_radius(G::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
 
 Return the injectivity radius for general complex unitary matrix manifolds, where the determinant is ``+1``,
 which is[^1]
@@ -670,7 +670,7 @@ which is[^1]
     \operatorname{inj}_{\mathrm{SU}(n)} = π \sqrt{2}.
 ```
 """
-function injectivity_radius(::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+function injectivity_radius(::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
     return π * sqrt(2.0)
 end
 
@@ -849,20 +849,20 @@ function manifold_dimension(M::GeneralUnitaryMatrices{<:Any,ℝ})
     return div(n * (n - 1), 2)
 end
 @doc raw"""
-    manifold_dimension(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+    manifold_dimension(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
 
 Return the dimension of the manifold of special unitary matrices.
 ```math
 \dim_{\mathrm{SU}(n)} = n^2-1.
 ```
 """
-function manifold_dimension(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+function manifold_dimension(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
     n = get_parameter(M.size)[1]
     return n^2 - 1
 end
 
 @doc raw"""
-    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℝ,AbsoluteDeterminantOneMatrices})
+    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℝ,AbsoluteDeterminantOneMatrixType})
 
 Volume of the manifold of real orthogonal matrices of absolute determinant one. The
 formula reads [BoyaSudarshanTilma:2003](@cite):
@@ -875,12 +875,14 @@ formula reads [BoyaSudarshanTilma:2003](@cite):
 \end{cases}
 ```
 """
-function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℝ,AbsoluteDeterminantOneMatrices})
+function manifold_volume(
+    M::GeneralUnitaryMatrices{<:Any,ℝ,AbsoluteDeterminantOneMatrixType},
+)
     n = get_parameter(M.size)[1]
-    return 2 * manifold_volume(GeneralUnitaryMatrices(n, ℝ, DeterminantOneMatrices))
+    return 2 * manifold_volume(GeneralUnitaryMatrices(n, ℝ, DeterminantOneMatrixType))
 end
 @doc raw"""
-    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℝ,DeterminantOneMatrices})
+    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℝ,DeterminantOneMatrixType})
 
 Volume of the manifold of real orthogonal matrices of determinant one. The
 formula reads [BoyaSudarshanTilma:2003](@cite):
@@ -896,7 +898,7 @@ formula reads [BoyaSudarshanTilma:2003](@cite):
 It differs from the paper by a factor of `sqrt(2)` due to a different choice of
 normalization.
 """
-function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℝ,DeterminantOneMatrices})
+function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℝ,DeterminantOneMatrixType})
     n = get_parameter(M.size)[1]
     vol = 1.0
     if n % 2 == 0
@@ -918,7 +920,7 @@ function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℝ,DeterminantOneMatri
     return vol
 end
 @doc raw"""
-    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℂ,AbsoluteDeterminantOneMatrices})
+    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℂ,AbsoluteDeterminantOneMatrixType})
 
 Volume of the manifold of complex general unitary matrices of absolute determinant one. The
 formula reads [BoyaSudarshanTilma:2003](@cite)
@@ -927,7 +929,9 @@ formula reads [BoyaSudarshanTilma:2003](@cite)
 \sqrt{n 2^{n+1}} π^{n(n+1)/2} \prod_{k=1}^{n-1}\frac{1}{k!}.
 ```
 """
-function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℂ,AbsoluteDeterminantOneMatrices})
+function manifold_volume(
+    M::GeneralUnitaryMatrices{<:Any,ℂ,AbsoluteDeterminantOneMatrixType},
+)
     n = get_parameter(M.size)[1]
     vol = sqrt(n * 2^(n + 1)) * π^(((n + 1) * n) // 2)
     kf = 1
@@ -938,7 +942,7 @@ function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℂ,AbsoluteDeterminant
     return vol
 end
 @doc raw"""
-    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+    manifold_volume(::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
 
 Volume of the manifold of complex general unitary matrices of determinant one. The formula
 reads [BoyaSudarshanTilma:2003](@cite)
@@ -947,7 +951,7 @@ reads [BoyaSudarshanTilma:2003](@cite)
 \sqrt{n 2^{n-1}} π^{(n-1)(n+2)/2} \prod_{k=1}^{n-1}\frac{1}{k!}.
 ```
 """
-function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrices})
+function manifold_volume(M::GeneralUnitaryMatrices{<:Any,ℂ,DeterminantOneMatrixType})
     n = get_parameter(M.size)[1]
     vol = sqrt(n * 2^(n - 1)) * π^(((n - 1) * (n + 2)) // 2)
     kf = 1
@@ -985,10 +989,10 @@ is
   \operatorname{proj}_{\mathrm{U}(n,𝔽)} \colon p ↦ U V^\mathrm{H}.
 ````
 """
-project(::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrices}, p) where {𝔽}
+project(::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrixType}, p) where {𝔽}
 
 function project!(
-    ::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrices},
+    ::GeneralUnitaryMatrices{<:Any,𝔽,AbsoluteDeterminantOneMatrixType},
     q,
     p,
 ) where {𝔽}
@@ -1080,7 +1084,7 @@ end
     riemann_tensor(::GeneralUnitaryMatrices, p, X, Y, Z)
 
 Compute the value of Riemann tensor on the [`GeneralUnitaryMatrices`](@ref) manifold.
-The formula reads [Rentmeesters:2011](@cite) 
+The formula reads [Rentmeesters:2011](@cite)
 
 ```math
 R(X,Y)Z=\frac{1}{4}[Z, [X, Y]].